refactor: move compat module
