+++ client/main.lua
@@ -24,6 +24,7 @@ end
 
 exports('GetVehiclesByCategory', GetVehiclesByCategory)
 
+---@deprecated use ox_inventory exports directly.
 ---@return table<number, Weapon>
 function GetWeapons()
     return QBX.Shared.Weapons
@@ -31,6 +32,7 @@ end
 
 exports('GetWeapons', GetWeapons)
 
+---@deprecated
 ---@return table<string, vector4>
 function GetLocations()
     return QBX.Shared.Locations
