+++ server/main.lua
@@ -39,6 +39,7 @@ end
 
 exports('GetVehiclesByCategory', GetVehiclesByCategory)
 
+---@deprecated use ox_inventory exports directly.
 ---@return table<number, Weapon>
 function GetWeapons()
     return QBX.Shared.Weapons
@@ -46,6 +47,7 @@ end
 
 exports('GetWeapons', GetWeapons)
 
+---@deprecated
 ---@return table<string, vector4>
 function GetLocations()
     return QBX.Shared.Locations
